refactored class and package names
