+++ README.md
@@ -1,5 +1,13 @@
 [![Build Status](https://travis-ci.org/abdulwahabO/identity-provider-demo.svg?branch=master)](https://travis-ci.org/abdulwahabO/identity-provider-demo)
 
+# About
+
+OAuth2 Passwordless login demo ?
+
+-- include how to build using a profile...
+
+todo --- put description of high level workings..
+
 # User Login Demo
 
 ![User Login Demo](demo.gif)
