+++ client/src/main/java/demo/oauth2passwordless/client/ClientApplication.java
@@ -1,0 +1,13 @@
+package demo.oauth2passwordless.client;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ClientApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ClientApplication.class, args);
+	}
+
+}
