--- client/src/main/java/me/oauth2providerdemo/client/ClientApplication.java
@@ -1,13 +1,0 @@
-package me.oauth2providerdemo.client;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ClientApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ClientApplication.class, args);
-	}
-
-}
