+++ pom.xml
@@ -2,13 +2,13 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>me.abdulwahabO</groupId>
-	<artifactId>oauth2-provider-demo</artifactId>
+	<groupId>demo.oauth2passwordless</groupId>
+	<artifactId>parent</artifactId>
 	<version>1.0.0-SNAPSHOT</version>
 	<packaging>pom</packaging>
-	<name>OAuth2 Provider Demo</name>
+	<name>Demo of a passwordless Oauth2 login flow</name>
 	<description>Authorization Server And Simple Web Client</description>
-
+	
 	<modules>
 		<module>server</module>
 		<module>client</module>
