+++ server/pom.xml
@@ -8,7 +8,7 @@
 		<version>2.1.1.RELEASE</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>me.oauth2providerdemo</groupId>
+	<groupId>demo.oauth2passwordless</groupId>
 	<artifactId>server</artifactId>
 	<version>1.0.0-SNAPSHOT</version>
 	<packaging>jar</packaging>
