+++ server/src/main/java/demo/oauth2passwordless/dto/UserDto.java
@@ -1,0 +1,41 @@
+package demo.oauth2passwordless.dto;
+
+public class UserDto {
+
+    private String name;
+    private String email;
+    private String phone;
+    private String appId;
+
+    public String getPhone() {
+        return phone;
+    }
+
+    public void setPhone(String phone) {
+        this.phone = phone;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getAppId() {
+        return appId;
+    }
+
+    public void setAppId(String appId) {
+        this.appId = appId;
+    }
+}
