+++ server/src/main/java/demo/oauth2passwordless/exception/AppCreationException.java
@@ -1,0 +1,8 @@
+package demo.oauth2passwordless.exception;
+
+public class AppCreationException extends Exception {
+
+    public AppCreationException(String message) {
+        super(message);
+    }
+}
