+++ server/src/main/java/demo/oauth2passwordless/exception/AuthenticationException.java
@@ -1,0 +1,8 @@
+package demo.oauth2passwordless.exception;
+
+public class AuthenticationException extends Exception {
+
+    public AuthenticationException(String message) {
+        super(message);
+    }
+}
