+++ server/src/main/java/demo/oauth2passwordless/exception/NoSuchAppException.java
@@ -1,0 +1,8 @@
+package demo.oauth2passwordless.exception;
+
+public class NoSuchAppException extends Exception {
+
+    public NoSuchAppException(String message) {
+        super(message);
+    }
+}
