+++ server/src/main/java/demo/oauth2passwordless/exception/NoSuchUserException.java
@@ -1,0 +1,8 @@
+package demo.oauth2passwordless.exception;
+
+public class NoSuchUserException extends Exception {
+
+    public NoSuchUserException(String message) {
+        super(message);
+    }
+}
