+++ server/src/main/java/demo/oauth2passwordless/message/EmailService.java
@@ -1,0 +1,17 @@
+package demo.oauth2passwordless.message;
+
+import demo.oauth2passwordless.model.User;
+
+/**
+ *
+ */
+public interface EmailService {
+
+    /**
+     *
+     * @param user
+     * @param verificationLink
+     * @return
+     */
+    boolean sendUserVerificationEmail(User user, String verificationLink);
+}
