+++ server/src/main/java/demo/oauth2passwordless/message/TextMessageService.java
@@ -1,0 +1,16 @@
+package demo.oauth2passwordless.message;
+
+/**
+ *
+ */
+public interface TextMessageService {
+
+    /**
+     *
+     * @param password
+     * @param number
+     * @return
+     */
+    boolean sendOneTimePassword(String password, String number);
+
+}
