--- server/src/main/java/me/identityprovider/AuthServerApplication.java
@@ -1,14 +1,0 @@
-package me.identityprovider;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
-
-@SpringBootApplication
-@EnableCaching
-public class AuthServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AuthServerApplication.class, args);
-	}
-}
