--- server/src/main/java/me/identityprovider/exception/AppCreationException.java
@@ -1,8 +1,0 @@
-package me.identityprovider.exception;
-
-public class AppCreationException extends Exception {
-
-    public AppCreationException(String message) {
-        super(message);
-    }
-}
