--- server/src/main/java/me/identityprovider/exception/AuthenticationException.java
@@ -1,8 +1,0 @@
-package me.identityprovider.exception;
-
-public class AuthenticationException extends Exception {
-
-    public AuthenticationException(String message) {
-        super(message);
-    }
-}
