--- server/src/main/java/me/identityprovider/exception/NoSuchAppException.java
@@ -1,8 +1,0 @@
-package me.identityprovider.exception;
-
-public class NoSuchAppException extends Exception {
-
-    public NoSuchAppException(String message) {
-        super(message);
-    }
-}
