--- server/src/main/java/me/identityprovider/exception/NoSuchUserException.java
@@ -1,8 +1,0 @@
-package me.identityprovider.exception;
-
-public class NoSuchUserException extends Exception {
-
-    public NoSuchUserException(String message) {
-        super(message);
-    }
-}
