--- server/src/main/java/me/identityprovider/message/TextSender.java
@@ -1,16 +1,0 @@
-package me.identityprovider.message;
-
-/**
- *
- */
-public interface TextSender {
-
-    /**
-     *
-     * @param password
-     * @param number
-     * @return
-     */
-    boolean sendOneTimePassword(String password, String number);
-
-}
