+++ server/src/test/java/demo/oauth2passwordless/service/LoginTest.java
@@ -1,0 +1,16 @@
+
+
+/*
+    generate OTP and mock cache to return it when sent back?
+
+    create user and apps, mock app reposository to return app for a particular Id
+
+    mock user repo to return a particular user too.
+
+    mock text message service to return true...as if OTP was sent.
+
+
+    READ: JUnit guides....
+
+
+*/
