+++ server/src/test/java/demo/oauth2passwordless/utils/SecurityUtilTest.java
@@ -1,0 +1,10 @@
+
+
+
+/*
+
+    Test JWT encoding.. Encode a JWT, den decode it and confirm claims and signature
+
+
+
+*/
